Began work on Platoon and adjusted includes Went through and changed all of the includes to be what they should be for the CMakeList (../include/FILE.h -> FILE.h). Additionally, started work on the Platoon class.
Originially I wanted to have Platoon be a simple vector or a struct,
but I think with the amount of information that is going to be stored
on the platoon level it looks like it will need to be a class at this
point. Perhaps in the future this can be reverted, and the information
stored within the Company class.